requirements.txt
# Core dependencies
requests>=2.31.0
PyGithub>=2.1.1
gitpython>=3.1.40

# LangChain for chunking and embeddings
langchain>=0.1.0
langchain-text-splitters>=0.0.1
langchain-community>=0.0.20  # For HuggingFace embeddings
langchain-ollama>=0.1.0  # For Ollama LLM integration
langchain-core>=0.1.0
langgraph>=0.0.1  # For workflow orchestration

# Embedding and Vector Store
qdrant-client>=1.7.0  # Qdrant vector database
tiktoken>=0.5.2
sentence-transformers>=2.2.0  # Required for HuggingFace BGE embeddings

# Utilities
python-dotenv>=1.0.0

# FastAPI for HITL web interface
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
typing-extensions>=4.9.0

# Development
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.12.0
ruff>=0.1.9
